+++ misc/config_tools/schema/checks/rdt_support.xsd
@@ -89,4 +89,12 @@ This error cannot be fixed by adjusting the configuration. Report a `GitHub issu
     </xs:annotation>
   </xs:assert>
 
+  <xs:assert test="every $needed in number-of-clos-id-needed(/acrn-config) satisfies
+                   every $capacity in min(//caches/cache/capability[@id='CAT']/clos_number) satisfies
+                   $needed &lt; 0">
+    <xs:annotation acrn:severity="error" acrn:report-on="/acrn-config/hv/CACHE_REGION">
+      <xs:documentation>The current CAT configuration requires {string($needed)} CLOS IDs, which exceeds the capacity of the platform which supports {string($capacity)} at most.</xs:documentation>
+    </xs:annotation>
+  </xs:assert>
+
 </xs:schema>
