+++ misc/config_tools/schema/datachecks.xsd
@@ -17,6 +17,7 @@
       <xi:include href="checks/virtio_devices.xsd" xpointer="xpointer(id('root')/*)" />
       <xi:include href="checks/vuart_config.xsd" xpointer="xpointer(id('root')/*)" />
       <xi:include href="checks/ivsh_memory.xsd" xpointer="xpointer(id('root')/*)" />
+      <xi:include href="checks/rdt_support.xsd" xpointer="xpointer(id('root')/*)" />
 
     </xs:complexType>
   </xs:element>
